-Historico de kahoots solo muestra las partidas terminadas - Control del unsubcribed, si se cierra la pestaña por parte del host o del participante y segun el estado del kahoot game se comporta de una forma u otra

+++ app/controllers/kahoot_participants_controller.rb
@@ -33,15 +33,18 @@ class KahootParticipantsController < ApplicationController
   end
 
   def destroy
-    session[:left_game] = true # Marca que el usuario salió manualmente
-    kahoot_game = @participant.kahoot_game
-    if @participant.destroy
-      show = kahoot_game.kahoot_participants.count > 0 ? "true" : "false"
-      KahootGameChannel.broadcast_to(kahoot_game.host, {type: "update_start_btn", show: show})
-      KahootGameChannel.broadcast_to(kahoot_game, { type: "player_left", user_id: current_user.id })
-      redirect_to new_kahoot_participant_path, notice: "Has salido de la partida."
+    if @participant
+      kahoot_game = @participant.kahoot_game
+      if @participant.destroy
+        show = kahoot_game.kahoot_participants.count > 0 ? "true" : "false"
+        KahootGameChannel.broadcast_to(kahoot_game.host, {type: "update_start_btn", show: show})
+        KahootGameChannel.broadcast_to(kahoot_game, { type: "player_left", user_id: current_user.id})
+        redirect_to new_kahoot_participant_path, notice: "Has salido de la partida."
+      else
+        redirect_to kahoot_game, alert: "No se pudo salir de la partida."
+      end
     else
-      redirect_to kahoot_game, alert: "No se pudo salir de la partida."
+      redirect_to new_kahoot_participant_path, notice: "Has salido de la partida."
     end
   end
